# cmd/r-vBackend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Risc-V-themed Web-Service",
        "title": "RVBACK",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/zzakku",
            "email": "nuhuh@lol.com"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/commands": {
            "get": {
                "description": "Получить все неудалённые команды. Доступно любому пользователю.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commands"
                ],
                "summary": "Получить все команды",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.successResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/commands/": {
            "post": {
                "description": "Добавить команду без изображения. Доступно модератору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commands"
                ],
                "summary": "Добавить команду",
                "parameters": [
                    {
                        "description": "Добавляемая команда",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Command"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.successResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Command": {
            "type": "object",
            "properties": {
                "com_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fmt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "rd_num": {
                    "type": "integer"
                },
                "rs_num": {
                    "type": "integer"
                }
            }
        },
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.successResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}